20251014_소재매칭_sp/requirements.txt
google-cloud-bigquery>=3.0.0
pandas>=1.5.0
db-dtypes>=1.0.0
